First TJ

+++ CRUD/Article/Article_read.php
@@ -67,7 +67,7 @@
 			  <td><?php echo $row['NumAngl']; ?></td>
 			  <td><?php echo $row['NumLang']; ?></td>
 			  <td><a href="Article_edit.php?id=<?php echo $row['NumArt'] ?>">Modifier</a></td>
-			  <td><a href="Article_delete.php?NumArt=<?php echo $row['NumArt'] ?>">Supprimer</a></td>
+			  <td><a href="Article_delete.php?NumArt=<?php echo $row['NumArt'] ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet article ?');">Supprimer</a></td>
 			</tr>
 
 		<?php
